backend/requirements.txt
# My Dolla $ign - Backend Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
flask==3.0.2
flask-cors==4.0.0

# OpenAI API for AI-powered analysis
openai==1.12.0

# Environment variables
python-dotenv==1.0.1

# Data validation
pydantic==2.6.1

# For production deployment (optional)
gunicorn==21.2.0
